fix(agent-adapter): canonicalize opencode history cwd filter; parallel read; shared test stream
